GFORMS-1452 - Implement Company Registration Number retrieval within form
