+++ app/uk/gov/hmrc/gform/gform/GformModule.scala
@@ -22,7 +22,6 @@ import cats.instances.future._
 import scala.concurrent.{ ExecutionContext, Future }
 import uk.gov.hmrc.gform.addresslookup.{ AddressLookupController, AddressLookupModule }
 import uk.gov.hmrc.gform.akka.AkkaModule
-import uk.gov.hmrc.gform.api.CompanyInformationAsyncConnector
 import uk.gov.hmrc.gform.auditing.AuditingModule
 import uk.gov.hmrc.gform.auth.{ AgentEnrolmentController, AuthModule, ErrorController }
 import uk.gov.hmrc.gform.bars.BankAccountReputationAsyncConnector
@@ -143,12 +142,6 @@ class GformModule(
   val bankAccountReputationConnector =
     new BankAccountReputationAsyncConnector(wSHttpModule.auditableWSHttp, barsBaseUrl)
 
-  private val companyInformationBaseUrl = s"${configModule.serviceConfig.baseUrl("company-information")}"
-  private val apiKey = s"${configModule.serviceConfig.getConfString("company-information.apiKey", "")}"
-
-  val companyInformationConnector =
-    new CompanyInformationAsyncConnector(wSHttpModule.auditableWSHttp, companyInformationBaseUrl, apiKey)
-
   val addToListProcessor = new FormProcessor(
     playBuiltInsModule.i18nSupport,
     processDataService,
@@ -159,7 +152,6 @@ class GformModule(
     fileUploadModule.fileUploadService,
     formControllerRequestHandler,
     bankAccountReputationConnector,
-    companyInformationConnector,
     addressLookupModule.addressLookupService
   )
